backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
databases[postgresql]==0.8.0
alembic==1.13.0
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# AWS Services
boto3==1.34.0
botocore==1.34.0

# AI/ML
openai==1.3.7
tiktoken==0.5.1

# Document Processing
PyPDF2==3.0.1
reportlab==4.0.7
Pillow==10.1.0

# Utilities
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Validation & Parsing
email-validator==2.1.0
phonenumbers==8.13.26
